requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
filelock==3.25.2
Flask==3.1.3
fonttools==4.62.1
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mysql-connector-python==9.3.0
networkx==3.6.1
numpy==2.3.1
opencv-python==4.13.0.92
packaging==26.0
passlib==1.7.4
pillow==11.2.1
polars==1.39.2
polars-runtime-32==1.39.2
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.23
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.42.0
Werkzeug==3.1.6
